Added a skeleton for the first XIA protocol
The first XIA protocol is `XIA_RAW'.

The code doesn't do anything besides issuing messages in dmesg,
the goal is to get a `working' template in which specific features
can be worked out.

The code isn't `working' yet, it still misses pieces such as
struct proto_ops, and network namespace.