Remove kfree() from free_fxid()
Having kfree() in free_fxid()'s releasing chain forces struct fib_xid
to always be allocated with kmalloc(), and forbids struct fib_xid be
just a sub-struct in a larger struct.

This constraint became troublesome for XDP principal, and had to be
refactored.

The code of the other principals were adjusted.