XIA/Serval: replace local headers to kernel ones
Serval's codebase keeps userland and kernel code intertwined,
and some headers are really meant to simplify the work of
userland code given that those headers are readily available
in the kernel.

Given that Linux XIA is only meant to be a native network
stack, this patch drops the easy to remove headers.