XIA/HID: implement efficient monitoring algorithm
Every struct hid_dev now has a timer to periodically begin a
monitoring interval. It is also necessary to allocate memory
for the hardware address of a temporary target for each
monitoring interval. This hardware address is under a lock
only when it is written-to by the single monitoring thread
and read by the potentially-numerous packet-handling threads.
Reads of this target hardware address in the monitoring
thread do not need to be locked.

There are also flags for whether monitoring is currently
occuring, for determining whether contact has been made with
any neighbor during the current monitoring interval (to
diagnose network partitions), for whether the neighbor has
yet been remonitored if failure is suspected, and for whether
the algorithm is currently in the investigative phase. These
flags are defined in net/xia/ppal_hid/nwp.c.

Upon arrival of a neighbor list, the monitoring algorithm may
begin in process_neigh_list(). monitor() initiates a monitoring
interval, using pick_random_neigh() to select a random target
to monitor.

To accomplish this, the monitoring algorithm first pings the
selected target, which may result in a timer expiration and
callback of end_ping_failure(). In that case, investigative
neighbors may be used to verify that the given target has
failed. If, at the end of the investigative time period, no
response from the target has been received, a timer expiry
callbacks into end_interval_failure(), which may declare a
node as failed.

A function find_update_neigh_sc() is added to update the status
of a hardware address for which we don't have a struct
hrdw_addr reference. If we're updating a node that has failed,
we need to estimate the remote clock time using the difference
between the old clock time and the new clock time. The list
replacement operation is made faster here by the RCU facility.

The forward declaration of end_interval_failure() is necessary
because of the cyclical nature between calls to monitor(),
end_ping_failure(), and end_interval_failure().

In order to send monitoring packets, send_monitoring() is
implemented. Note that const u8 *inv should not be used
for Ping and Ack packets, and in those cases should be set to
NULL.

nwp_rcv() may now accept monitoring packets and distribute
them to process_monitoring(), which updates clocks, changes
timers, and dispatches packets as necessary depending on the
type of packet that has been received. When a monitoring
iteration is successful (an Ack is received), the monitoring
interval will end in end_interval_success().