XIA/Serval: fix handling of 3-way handshake's ACK
The serval request socket corresponding to the ACK wasn't being
found because the code was using the source address instead of
the destination address to find the corresponding request socket.

Moreover, serval_sal_request_sock_handle() was not testing
the peer's nonce correctly, what was rendering the ACK packet as
invalid.