xia/u4id: allow a remote U4ID XID to be a sink
It doesn't make sense to send packets to an XIP address in which
the final sink is a remote U4ID XID because no transport is expect
to deal with these packets at the destination.
Nevertheless, if the sink condition comes from a routing redirect,
it is fine since that the U4ID is just scoping the sink in this case.